+++ deye-solarman-diagnostics/rootfs/usr/src/app/deye_solarman_diagnostics/__main__.py
@@ -3,4 +3,4 @@ from .solarman import SolarmanClient
 
 
 if __name__ == "__main__":
-	main(SolarmanClient)
+	main(SolarmanClient,source="solarman_tcp",source_name="Deye Solarman Local")